requirements.txt
flask==3.0.3
python-dotenv==1.0.1
openpyxl==3.1.5
httpx==0.27.2
pytest==8.3.3
